+++ pytest-profiling/tests/integration/profile/tests/unit/test_regex.py
@@ -1,0 +1,6 @@
+import os
+
+
+def test_regex():
+    os.makedirs("bar", exist_ok=True)
+    os.chdir("bar")